+++ MemoryPuzzle/MemoryPuzzle/GameMode.swift
@@ -1,0 +1,9 @@
+import UIKit
+
+extension ViewController{
+    
+    @IBAction func gameModeAction(_ sender: Any) {
+        
+        
+    }
+}
