+++ MemoryPuzzle/MemoryPuzzle/Reset.swift
@@ -1,0 +1,12 @@
+import UIKit
+
+extension ViewController{
+
+    @IBAction func resetAction(_ sender: Any) {
+        
+        
+        
+    }
+    
+
+}
